PLAT-224: 1. move Marshaler to insolar.payload. 2. add comment to magic number. 3. propagate context to adapter creator
