+++ ledger-core/v2/insolar/payload/marshaler.go
@@ -3,7 +3,7 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/insolar/blob/master/LICENSE.md.
 
-package insolar
+package payload
 
 type Marshaler interface {
 	Marshal() ([]byte, error)
